fix: backtick processing
The current source generation produces hex encoded backticks, resulting in
invalid code.

There seems to be no way to properly encode or escape them, as the final code is
itself part of a backticked multi-lines string.

Go does not allow nesting or escaping backticks. There's some proposals to address this issue:
- https://github.com/golang/go/issues/32190
- https://github.com/golang/go/issues/32590

but look like nothing got accepted there yet.

The current sources have been updated to use string concatenation
instead of backticks, and the generator got updated to panic if it
encounter any.

This is not an ideal solution, as some sources may require the usage of
backticks at some point, but it would probably require a deeper rework
of the generation to fix it (maybe external files + go:embed ?)